backend/requirements.txt
fastapi
uvicorn
pydantic[email]
passlib[bcrypt]
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
colorama==0.4.6
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
greenlet==3.1.1 
sqlalchemy==2.0.36
asyncpg==0.30.0